reject changes to this file
If we change any of these routes to ES6, we should change all. For the time being, let's keep them as they are to ensure everyone working on them understands the syntax.

Also, the changes to lines 48-49 are not acceptable changes: the front end is passing back the full idea object, not just the id and this path needs to be used for several different kinds of updates.