reject changes
please do not make changes to variable names without speaking to the code author--Johnny says this id's must be all lowercase, not camel case

In general, let's refrain from making unnecessary changes to code that others are working with. Keep the style that the code author is used to seeing until we get closer to the end.